requirements.txt
python-decouple==3.6
holidays==0.13